move db context to Data
