+++ Surabhi-GeekHunter/GeekHunter/GeekHunter/Controllers/CandidatesController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using GeekHunter.Models;
+using GeekHunter.Data;
 
 namespace GeekHunter.Controllers
 {
