+++ Surabhi-GeekHunter/GeekHunter/GeekHunter/Data/IDbContext.cs
@@ -3,8 +3,9 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using GeekHunter.Models;
 
-namespace GeekHunter.Models
+namespace GeekHunter.Data
 {
     interface IDbContext : IDisposable
     {
